Migrate from MediaPlayer to ExoPlayer

+++ OpenEdXMobile/src/main/java/org/edx/mobile/model/VideoModel.java
@@ -47,7 +47,7 @@ public interface VideoModel {
 
     String getSectionName();
 
-    int getLastPlayedOffset();
+    long getLastPlayedOffset();
 
     String getLmsUrl();
